Cleaned up code, added concern for post and community models
